CS631/stuff/app/requirements.txt
Flask
Flask-MySQL
mysql-connector
simplejson
flask_wtf
email_validator
Flask-Cache
python-dotenv